pyproject.toml
[tool]
[tool.poetry]
name = "edtools"
version = "1.0.4"
description = "Collection of tools for automated processing and clustering of electron diffraction data."
keywords = [
	"electron-diffraction",
	"microed",
	"xds",
	"pipeline",
	"cluster-analysis",
]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"License :: OSI Approved :: BSD License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
]
homepage = "http://github.com/instamatic-dev/edtools"
repository = "http://github.com/instamatic-dev/edtools"
documentation = "http://github.com/instamatic-dev/edtools"
maintainers = ["Stef Smeets <s.smeets@esciencecenter.nl>"]
authors = ["Stef Smeets <s.smeets@esciencecenter.nl>"]
readme = "readme.md"
license = "BSD-3-clause"
include = [
    ".pre-commit.yaml",
    "LICENCE",
    "setup.py",
]

[tool.poetry.dependencies]
python = ">=3.8.0"
numpy = ">=1.18.2"
matplotlib = ">=3.2.1"
scipy = ">=1.4.1"
pandas = ">=1.0.3"
uncertainties = ">=3.1.2"
PyYAML = ">=5.4.1"
openpyxl = ">=3.0.10"

[tool.poetry.dev-dependencies]
check-manifest = "*"
pre-commit = "*"
bump2version = "^1.0.1"

[tool.poetry.scripts]
"edtools.autoindex"           = "edtools.autoindex:main"
"edtools.cluster"             = "edtools.cluster:main"
"edtools.find_cell"           = "edtools.find_cell:main"
"edtools.extract_xds_info"    = "edtools.extract_xds_info:main"
"edtools.run_pointless"       = "edtools.run_pointless:main"
"edtools.make_xscale"         = "edtools.make_xscale:main"
"edtools.make_shelx"          = "edtools.make_shelx:main"
"edtools.update_xds"          = "edtools.update_xds:main"
"edtools.find_rotation_axis"  = "edtools.find_rotation_axis:main"
"edtools.find_beam_center"  = "edtools.find_beam_center:main"

[tool.poetry.urls]
"Bug Reports" = "https://github.com/instamatic-dev/edtools/issues"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}
versioning = "semver"

[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"
